Some functionality addition to clone_func():
Added a flag parameterIsExecutable which switches between whether parameter received from virgo_clone or telnet is:
 1. name of a function (or)
 2. name of a binary

Linux kernel has call_usermodexxx() functions that do upcall to userspace and execute a binary. Presently virgo_clone() is only
for creating a cloned thread in remote cloud node and sends name of a function. To support kernel upcall, based on the flag 
parameterIsExecutable call_usermodehelper() is invoked on the received data which has name of the binary.

If parameterIsExecutable is set to 1 the data received from virgo_clone() is not a function but name of executable
This executable is then run on usermode using call_usermodehelper() which internally takes care of queueing the workstruct
and executes the binary in userspace as child of keventd and reaps silently. Thus workqueue component of kernel is indirectly made use of.
This is sometimes more flexible alternative that executes a binary itself on cloud and 
is preferable to clone()ing a function on cloud. Virgo_clone() syscall client or telnet needs to send the message with name of binary.
If parameterIsExecutable is set to 0 then data received from virgo_clone() is name of a function.

- Ka.Shrinivaasan (alias) Shrinivas Kannan (alias) Srinivasan Kannan
  (http://sites.google.com/site/kuja27)