Major rewrite of virgo_cloudexec for output redirection with fd_install:
1. Opened a VFS stdout file called file_stdout and added fd_install for stdout and stderr to that file so that kernel_execve() inherits
these file descriptors and goes upto usermode and all printf()s in usermode executable or functions get logged into this file created by
Kernel. This was earlier difficult due to lack of stdout fd_install in kernel 3.x and seemed to require a copy_from_user() or modification
of kmod.c or netlink listeners. Name of this Kernel-User communication file is hardcoded as virgo_cloudexec_upcall_usermode_log.txt. Lot of testing
was carried out for parameterIsExecutable=0 and parameterIsExecutable=1 with both virgo_clone test case and also telnet. All output logs
have been committed under test_logs/. This is a significant feature addition that makes Kernel To Userspace communication and output redirection
possible which was muted earlier. Though this may be viewed as a security risk, in the hindsight it is a serious feature that enables
Kernel To Usermode logging possible without which logging within user mode was earlier impossible and there was no way to know what user has
executed except by return code which is tedious to decrypt.

2.Important thing is that fd_install code is within virgo module and not within kmod.c. 

3.All bash shell invocations have been dropped and executables are invoked directly. filp_open+fd_install+filp_close code executes now for
each invocation of usermode helper which can be in module init() and exit() as an optimization.
 This was tested but for some reason it didn't work but should have worked.