Initial Prototype code for VIRGO Memory Pooling feature (virgo_malloc() etc.,). This code is mostly copied from cpupooling directory and changed for Memory Pooling All the basic functionalities like 2 user modes and Kernel modes execution, multithreaded execution and code organization remain the same.
New header has been added in include/linux for VIRGO mempool kernel service - virgocloudexecmempoolsvc.h and virgo.h has been modified
to include mempool related virgo_ops declarations.Only the basic VIRGO mempool driver build works in this commit.

This is not tested code as complete testing can be done only after virgo_malloc() syscall is implemented fully which is a work for few months.
At present a not-built, not-tested pseudocode kind of virgo_malloc() is committed under virgo_malloc() for reference which will be 
overridden in future commits.That would be another replication of virgo_clone() and suitably modified for virgo_malloc() and kernel 
 builds for new syscalls take considerable time.